Working on Vege editor integration
